+++ lab1/task3-feistel/test/FeistelCipherTest.java
@@ -75,7 +75,7 @@ public class FeistelCipherTest {
         System.out.println("Тест 2: Разное количество раундов");
         try {
             byte[] key = {0x01, 0x02, 0x03, 0x04};
-            byte[] plaintext = {0x12, 0x34, 0x56, 0x78, 0x9A, 0xBC, 0xDE, 0xF0};
+            byte[] plaintext = {0x12, 0x34, 0x56, 0x78, (byte)0x9A, (byte)0xBC, (byte)0xDE, (byte)0xF0};
 
             int[] roundCounts = {2, 4, 8, 16};
